--- a/.cursor/rules/cloudformation-linting.mdc
+++ b/.cursor/rules/cloudformation-linting.mdc
@@ -1,40 +1,29 @@
----
-description: CloudFormation-specific linting rules
-globs: ["**/*.yaml", "**/*.yml", "**/*.template"]
-alwaysApply: false
----
 
-# CloudFormation Linting
+# CloudFormation Linting Rules
 
-## BANNED: Generic YAML Linters for CloudFormation
-- NEVER use generic YAML linters on CloudFormation templates
-- NEVER use tools that don't understand `!Sub`, `!Ref`, `!GetAtt`
-- NEVER use tools that flag CloudFormation intrinsic functions as errors
+## File Detection
+- Apply to CloudFormation templates and infrastructure YAML files
+- Focus on AWS infrastructure patterns
+- Include models directory for infrastructure templates
 
-## REQUIRED: CloudFormation-Specific Tools
+## Validation Requirements
+- Validate CloudFormation syntax
+- Check for security best practices
+- Ensure proper resource naming conventions
+- Validate IAM permissions follow least privilege
 
-### Primary Tool: cfn-lint
-```bash
-# Install
-pip install cfn-lint
+## Common Issues to Catch
+- Hardcoded credentials or secrets
+- Overly permissive IAM policies
+- Missing required properties
+- Invalid resource references
+- Security group misconfigurations
 
-# Validate CloudFormation templates
-cfn-lint models/Openflow-Playground.yaml
-
-# Validate with specific rules
-cfn-lint --append-rules E3002,E3003 models/Openflow-Playground.yaml
-```
-
-### Alternative Tools
-- **AWS CloudFormation Designer** - Visual validation
-- **AWS Toolkit for VS Code** - IntelliSense and validation
-- **AWS SAM CLI** - For serverless templates
-- **CloudFormation Guard** - Policy validation
-
-## File Patterns to Ignore in Generic Linters
-- `*.template.yaml`
-- `*.template.yml` 
-- `*cloudformation*.yaml`
+## Best Practices
+- Use parameters for configurable values
+- Implement proper tagging strategies
+- Follow AWS Well-Architected Framework
+- Use CloudFormation macros for DRY principles
 - `*cloudformation*.yml`
 - `models/Openflow-Playground.yaml`