--- a/.cursor/rules/deterministic-editing.mdc
+++ b/.cursor/rules/deterministic-editing.mdc
@@ -1,6 +1,6 @@
 ---
 description: Use deterministic tools for file editing
-globs: ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.toml", "**/*.ini", "**/*.cfg", "**/*.mdc", "**/*.py"]
+globs: ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.toml", "**/*.ini", "**/*.cfg", "**/*.mdc", "**/*.py", "**/*.xml", "**/*.properties", "**/*.env"]
 alwaysApply: true
 ---