--- a/.cursor/rules/security-credentials.mdc
+++ b/.cursor/rules/security-credentials.mdc
@@ -1,6 +1,9 @@
 ---
-title: Security: Credentials & Sensitive Data
+description: Security rules for credentials and sensitive data
+globs: ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.env", "**/*.sh", "**/*.py", "**/*.md", "**/*.txt"]
+alwaysApply: true
 ---
+
 
 # Security: Credentials & Sensitive Data