--- a/.cursor/rules/security-credentials.mdc
+++ b/.cursor/rules/security-credentials.mdc
@@ -1,6 +1,6 @@
 ---
 description: Security rules for credentials and sensitive data
-globs: ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.env", "**/*.sh", "**/*.py", "**/*.md", "**/*.txt"]
+globs: ["**/*.yaml", "**/*.yml", "**/*.json", "**/*.env*", "**/*.sh", "**/*.py", "**/*.md", "**/*.txt", "**/*.properties", "**/*.ini", "**/*.cfg", "**/*.toml"]
 alwaysApply: true
 ---