--- a/.cursor/rules/yaml-type-specific.mdc
+++ b/.cursor/rules/yaml-type-specific.mdc
@@ -1,44 +1,73 @@
 # YAML File Type-Specific Rules
 
-## File Type Detection and Routing
+## Core Principle: Heuristic Evaluation Over Enumeration
 
-### 1. Infrastructure/CloudFormation YAML
-**Patterns:** `*cloudformation*.yaml`, `*infrastructure*.yaml`, `*aws*.yaml`, `models/*.yaml`
-**Apply:** CloudFormation-specific linting rules
-**Tools:** cfn-lint, AWS CloudFormation Designer
-**Validation:** CloudFormation syntax, AWS resource validation
+**When encountering YAML files, evaluate their purpose and apply appropriate rules rather than trying to predict every possible format quirk.**
 
-### 2. Configuration YAML
-**Patterns:** `config*.yaml`, `settings*.yaml`, `*.config.yaml`
-**Apply:** Configuration validation rules
-**Tools:** jsonschema, yamllint
-**Validation:** Schema validation, required fields, data types
+## Evaluation Strategy
 
-### 3. CI/CD Pipeline YAML
-**Patterns:** `.github/*.yaml`, `.gitlab-ci.yml`, `*.workflow.yaml`, `azure-pipelines*.yml`
-**Apply:** CI/CD pipeline rules
-**Tools:** GitHub Actions validator, GitLab CI validator
-**Validation:** Workflow syntax, security practices
+### 1. File Purpose Detection
+**Ask yourself: "What is this YAML file trying to accomplish?"**
 
-### 4. Kubernetes YAML
-**Patterns:** `k8s/*.yaml`, `kubernetes/*.yaml`, `*.k8s.yaml`
-**Apply:** Kubernetes-specific rules
-**Tools:** kubectl validate, kubeval
-**Validation:** Kubernetes API validation, RBAC rules
+- **Infrastructure/CloudFormation**: AWS resources, CloudFormation templates
+- **Configuration**: App settings, environment configs, feature flags
+- **CI/CD**: GitHub Actions, GitLab CI, Azure Pipelines, Jenkins
+- **Container Orchestration**: Kubernetes, Docker Compose, Helm charts
+- **Application Config**: App-specific settings, database configs
+- **Schema/API**: OpenAPI specs, GraphQL schemas, data models
 
-### 5. Docker Compose YAML
-**Patterns:** `docker-compose*.yml`, `compose*.yml`
-**Apply:** Docker Compose rules
-**Tools:** docker-compose config
-**Validation:** Service definitions, network configuration
+### 2. Tool Selection by Purpose
+**Choose validation tools based on the file's purpose, not just its extension:**
 
-### 6. Application Configuration YAML
-**Patterns:** `app*.yaml`, `application*.yaml`, `*.app.yaml`
-**Apply:** Application config rules
-**Tools:** yamllint, custom schema validation
-**Validation:** Environment-specific configs, feature flags
+```bash
+# Infrastructure files
+cfn-lint *.yaml                    # CloudFormation
+kubectl validate -f *.yaml         # Kubernetes
+docker-compose config              # Docker Compose
 
-## Universal YAML Rules
+# Configuration files
+yamllint *.yaml                    # General YAML
+jsonschema -i schema.json *.yaml   # Schema validation
+
+# CI/CD files
+# GitHub Actions: Built-in validation
+# GitLab CI: gitlab-ci-lint
+# Azure Pipelines: az pipelines validate
+```
+
+### 3. When You Hit a New YAML Quirk
+
+**Follow this heuristic process:**
+
+1. **Identify the quirk**: What's different about this YAML format?
+2. **Determine the purpose**: What is this file trying to accomplish?
+3. **Find the right tool**: What tool understands this format?
+4. **Apply appropriate validation**: Use domain-specific tools, not generic YAML linters
+5. **Document the pattern**: Add to this rule for future reference
+
+### 4. Common YAML Quirks and Solutions
+
+#### CloudFormation Intrinsic Functions
+**Problem**: Generic YAML linters flag `!Ref`, `!Sub`, `!GetAtt` as errors
+**Solution**: Use `cfn-lint` instead of generic YAML validators
+
+#### Kubernetes Custom Resources
+**Problem**: Generic validators don't understand CRDs
+**Solution**: Use `kubectl validate` with proper API server
+
+#### GitHub Actions Context
+**Problem**: Generic YAML linters don't understand `${{ }}` expressions
+**Solution**: Use GitHub's built-in validation or `actionlint`
+
+#### Helm Templates
+**Problem**: Generic linters flag `{{ }}` as invalid YAML
+**Solution**: Use `helm template --validate` or `helm lint`
+
+#### Ansible Playbooks
+**Problem**: Generic linters don't understand Ansible-specific syntax
+**Solution**: Use `ansible-lint` for playbook validation
+
+## Universal YAML Principles
 
 ### Formatting Standards
 - Use 2-space indentation (never tabs)
@@ -53,16 +82,20 @@
 - Follow principle of least privilege
 
 ### Validation Strategy
-1. **Detect file type** based on path and content
-2. **Apply type-specific rules** and tools
-3. **Validate against appropriate schema**
-4. **Check security and best practices**
-5. **Ensure proper formatting and structure**
+1. **Detect file purpose** based on path, content, and context
+2. **Choose appropriate tools** for that purpose
+3. **Apply domain-specific validation** rather than generic YAML linting
+4. **Check security and best practices** for that domain
+5. **Document new patterns** when discovered
 
-## Common YAML Issues to Prevent
-- Mixed indentation (spaces vs tabs)
-- Incorrect YAML syntax (missing colons, wrong quotes)
-- Invalid references or anchors
-- Security vulnerabilities (hardcoded secrets)
-- Schema violations
-- Inconsistent formatting
+## When You Encounter New YAML Formats
+
+**Remember: The goal is not to predict every quirk, but to have a principled approach for handling them.**
+
+1. **Don't panic** - YAML quirks are common and expected
+2. **Identify the purpose** - What domain does this YAML serve?
+3. **Find the right tool** - Use domain-specific validators
+4. **Apply appropriate rules** - Don't force generic YAML linting
+5. **Document the pattern** - Add to this rule for future reference
+
+**The principle: Use the right tool for the job, not a generic tool for everything.**